refactor(inference): split openai_compat.rs into directory module under 500-line guideline
The single openai_compat.rs had grown to 785 lines, over the project's
500-line file guideline. Convert it to a directory module mirroring the
in-crate qnn::backend precedent:

- openai_compat/mod.rs   — implementation (393 lines)
- openai_compat/tests.rs — the #[cfg(test)] mod tests body (392 lines),
  declared via `#[cfg(test)] mod tests;` from mod.rs

Pure move, no behaviour change: the two `super`/`super::super` references
resolve identically after the split (tests' parent is still the
openai_compat module; classify_tests' super::super still reaches it). All
30 openai_compat tests retained and passing; public API unchanged.

Verified from src/ under the pinned 1.88 toolchain: fmt --check clean,
clippy --workspace -D warnings clean, full workspace test suite 0 failures.
